Attachment download: prefer IDs to FileObjects, more shared code
- Move shareable bits into lib-common (blob download & attachment-to-file conversion)
- Prefer UUIDs to FileObjects in shared APIs
    - FileObject smells a bit in its current form (the same data is transformed back and forth between it and ApplicationAttachment) and there's no real reason for the API calls to have anything more than an ID
