+++ frontend/packages/citizen-frontend/src/applications/editor/verification/ServiceNeedAttachments.tsx
@@ -6,13 +6,13 @@ import React, { useState } from 'react'
 import styled from 'styled-components'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
+import { attachmentToFile } from '@evaka/lib-common/src/utils/file'
 import FileDownloadButton from '@evaka/lib-components/src/molecules/FileDownloadButton'
 import InfoModal from '@evaka/lib-components/src/molecules/modals/InfoModal'
 import { Label } from '@evaka/lib-components/src/typography'
 import { Gap } from '@evaka/lib-components/src/white-space'
 import { faFile, faInfo } from '@evaka/lib-icons'
 import { ApplicationFormData } from '../../../applications/editor/ApplicationFormData'
-import { attachmentToFile } from '../../../applications/editor/FileUpload'
 import { useTranslation } from '../../../localization'
 import { getFileAvailability, getFileBlob } from '../../../applications/api'
